Each Flux dependency (Store, Selector) should be watched as Identifier (#24)
Since we have established rule that from `selectors` there should be imported only Selectors and as a default export imported from `stores` there has to be Store. There is no need to make the logic so complicated, taking into account that current severity of this rule is warning. Once the watched identifier is used, we should analyze it. 

Fixes #23